commit in the day 22.03.2024:20-45

+++ src/main/java/com/radik/my/project/services/UserService.java
@@ -8,6 +8,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -18,23 +19,26 @@ public class UserService {
 
     private final PasswordEncoder encoder;
 
-    public Optional<User> get(Long id) {
+    public User get(Long id) {
         if (id == null) throw new NotCorrectUserDetailsException("id не может быть null");
-        return userRepository.findById(id);
+        return userRepository.findById(id).orElse(null);
     }
 
-    public Optional<User> get(String email) {
+    public User get(String email) {
         if (email == null) throw new NotCorrectUserDetailsException("email не может быть null");
         return userRepository.findByEmail(email);
     }
 
     public void delete(Long id) {
         if (id == null) throw new NotCorrectUserDetailsException("id не может быть null");
+        if (!userRepository.existsById(id)) throw new NotCorrectUserDetailsException("пользователя с таким id не существует");
+
         userRepository.deleteById(id);
     }
 
     public User create(User user) {
         if (user == null) throw new NotCorrectUserDetailsException("user не может быть null");
+        if (ifExistEmail(user.getEmail())) return null;
 
         String passwordEncode = encoder.encode(user.getPassword());
         user.setPassword(passwordEncode);
@@ -45,13 +49,15 @@ public class UserService {
     public User update(User user) {
         if (user == null) throw new NotCorrectUserDetailsException("при обновление user не может быть null");
 
-        User oldUser = get(user.getId()).orElse(null);
-        if (oldUser == null) throw new NotCorrectUserDetailsException("обновление данных user -> такого user не существует");
+        if (!userRepository.existsById(user.getId())) throw new NotCorrectUserDetailsException("обновление данных user -> такого user не существует");
 
         return userRepository.save(user);
     }
 
     public User addCount(Long id, int count) {
+        if (count < 0 || id == null) throw new RuntimeException("некорректные данные");
+        if (!ifExistId(id)) throw new NotCorrectUserDetailsException("user с таким id не существует");
+
         Optional<User> optionalUser = userRepository.findById(id);
         if (optionalUser.isPresent()) {
             if (count == 0) return optionalUser.get();
@@ -65,5 +71,39 @@ public class UserService {
         } else throw new NotCorrectUserDetailsException("user с таким email не существует");
     }
 
+    public void modifyName(Long id, String name) {
+        if (id == null || Objects.isNull(name) || name.trim().isEmpty()) throw new RuntimeException("некорректные данные");
+        if (!userRepository.existsById(id)) throw new NotCorrectUserDetailsException("user с таким id не существует");
+
+        User user = userRepository.findById(id).get();
+        user.setFirstName(name);
+        userRepository.save(user);
+    }
+    public void modifySurname(Long id, String surname) {
+        if (id == null || Objects.isNull(surname) || surname.trim().isEmpty()) throw new RuntimeException("некорректные данные");
+        if (!userRepository.existsById(id)) throw new NotCorrectUserDetailsException("user с таким id не существует");
+
+        User user = userRepository.findById(id).get();
+        user.setLastName(surname);
+        userRepository.save(user);
+    }
+
+    public void modifyEmail(Long id, String email) {
+        if (id == null || Objects.isNull(email) || email.trim().isEmpty()) throw new RuntimeException("некорректные данные");
+        if (!userRepository.existsById(id)) throw new NotCorrectUserDetailsException("user с таким id не существует");
+        if (ifExistEmail(email)) throw new NotCorrectUserDetailsException("user с таким email уже существует");
+
+        User user = userRepository.findById(id).get();
+        user.setEmail(email);
+        userRepository.save(user);
+    }
+
+    public boolean ifExistEmail(String email) {
+        return userRepository.existsByEmail(email);
+    }
+
+    public boolean ifExistId(Long id) {
+        return userRepository.existsById(id);
+    }
 
 }
